Consolidate zurg's object-like utils into one file
