+++ packages/core-utilities/zurg/src/builders/object-like/index.ts
@@ -1,3 +1,2 @@
-export { getObjectLikeUtils } from "./getObjectLikeUtils";
+export { getObjectLikeUtils, withProperties } from "./getObjectLikeUtils";
 export { OBJECT_LIKE_BRAND, type BaseObjectLikeSchema, type ObjectLikeSchema, type ObjectLikeUtils } from "./types";
-export { withProperties } from "./withProperties";
